requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.5
sqlalchemy==2.0.36
pydantic==2.10.5
pydantic-settings==2.7.1
python-multipart==0.0.20
Pillow==11.1.0
exifread==3.0.0
httpx==0.28.1
python-dotenv==1.0.1
pytest==8.3.4
